arguments reading refactoring
