+++ src/main.rs
@@ -1,6 +1,7 @@
 use std::env;
-use std::fs;
 use std::process;
+
+use comand_line_program::Config;
 
 fn main() {
     let args: Vec<String> = env::args().collect();
@@ -13,25 +14,8 @@ fn main() {
     println!("Searching for {}", config.query);
     println!("In file {}", config.file_path);
 
-    let contents =
-        fs::read_to_string(config.file_path).expect("Should have been able to read the file");
-
-    println!("With text:\n{contents}");
-}
-
-struct Config {
-    query: String,
-    file_path: String,
-}
-
-impl Config {
-    fn build(args: &[String]) -> Result<Config, &'static str> {
-        if args.len() < 3 {
-            return Err("not enough arguments");
-        }
-        let query = args[1].clone();
-        let file_path = args[2].clone();
-
-        Ok(Config { query, file_path })
+    if let Err(e) = comand_line_program::run(config) {
+        println!("Application error: {e}");
+        process::exit(1);
     }
 }
